% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{BatchReport.outputs}
\alias{BatchReport.outputs}
\title{BatchReport.outputs Object}
\usage{
BatchReport.outputs()
}
\value{
BatchReport.outputs object
}
\description{
BatchReport.outputs Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other BatchReport functions: \code{\link{BatchReport.timeSpan}},
  \code{\link{BatchReport}}
}

