% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{BatchReport.timeSpan}
\alias{BatchReport.timeSpan}
\title{BatchReport.timeSpan Object}
\usage{
BatchReport.timeSpan(endTime = NULL, startTime = NULL)
}
\arguments{
\item{endTime}{End of the period included in the report}

\item{startTime}{Start of the period included in the report}
}
\value{
BatchReport.timeSpan object
}
\description{
BatchReport.timeSpan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Period included in the report. For reports containing all entities endTime is not set. Both startTime and endTime are inclusive.
}
\seealso{
Other BatchReport functions: \code{\link{BatchReport.outputs}},
  \code{\link{BatchReport}}
}

