% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{BatchReportList}
\alias{BatchReportList}
\title{BatchReportList Object}
\usage{
BatchReportList(items = NULL)
}
\arguments{
\item{items}{A list of batchReport resources that match the request criteria}
}
\value{
BatchReportList object
}
\description{
BatchReportList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A paginated list of batchReport resources returned in response to a youtubeAnalytics.batchReport.list request.
}

