% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{Group}
\alias{Group}
\title{YouTube Analytics API Objects 
Retrieves your YouTube Analytics data.}
\usage{
Group(Group.contentDetails = NULL, Group.snippet = NULL,
  contentDetails = NULL, etag = NULL, id = NULL, snippet = NULL)
}
\arguments{
\item{Group.contentDetails}{The \link{Group.contentDetails} object or list of objects}

\item{Group.snippet}{The \link{Group.snippet} object or list of objects}

\item{contentDetails}{No description}

\item{etag}{No description}

\item{id}{No description}

\item{snippet}{No description}
}
\value{
Group object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:23:49
filename: /Users/mark/dev/R/autoGoogleAPI/googleyoutubeAnalyticsv1beta1.auto/R/youtubeAnalytics_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Group Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Group functions: \code{\link{Group.contentDetails}},
  \code{\link{Group.snippet}}, \code{\link{groups.insert}},
  \code{\link{groups.update}}
}
