% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupItemListResponse}
\alias{GroupItemListResponse}
\title{GroupItemListResponse Object}
\usage{
GroupItemListResponse(etag = NULL, items = NULL)
}
\arguments{
\item{etag}{No description}

\item{items}{No description}
}
\value{
GroupItemListResponse object
}
\description{
GroupItemListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A paginated list of grouList resources returned in response to a youtubeAnalytics.groupApi.list request.
}

