% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\name{groupItems.insert}
\alias{groupItems.insert}
\title{Creates a group item.}
\usage{
groupItems.insert(GroupItem, onBehalfOfContentOwner = NULL)
}
\arguments{
\item{GroupItem}{The \link{GroupItem} object to pass to this method}

\item{onBehalfOfContentOwner}{Note: This parameter is intended exclusively for YouTube content partners}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtubepartner
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtubepartner)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/youtube/analytics/}{Google Documentation}

Other GroupItem functions: \code{\link{GroupItem.resource}},
  \code{\link{GroupItem}}
}

