% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{GroupListResponse}
\alias{GroupListResponse}
\title{GroupListResponse Object}
\usage{
GroupListResponse(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{No description}

\item{items}{No description}

\item{nextPageToken}{No description}
}
\value{
GroupListResponse object
}
\description{
GroupListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A paginated list of grouList resources returned in response to a youtubeAnalytics.groupApi.list request.
}

