% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_functions.R
\name{reports.query}
\alias{reports.query}
\title{Retrieve your YouTube Analytics reports.}
\usage{
reports.query(ids, start.date, end.date, metrics, currency = NULL,
  dimensions = NULL, filters = NULL, max.results = NULL, sort = NULL,
  start.index = NULL)
}
\arguments{
\item{ids}{Identifies the YouTube channel or content owner for which you are retrieving YouTube Analytics data}

\item{start.date}{The start date for fetching YouTube Analytics data}

\item{end.date}{The end date for fetching YouTube Analytics data}

\item{metrics}{A comma-separated list of YouTube Analytics metrics, such as views or likes,dislikes}

\item{currency}{The currency to which financial metrics should be converted}

\item{dimensions}{A comma-separated list of YouTube Analytics dimensions, such as views or ageGroup,gender}

\item{filters}{A list of filters that should be applied when retrieving YouTube Analytics data}

\item{max.results}{The maximum number of rows to include in the response}

\item{sort}{A comma-separated list of dimensions or metrics that determine the sort order for YouTube Analytics data}

\item{start.index}{An index of the first entity to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/youtube
\item https://www.googleapis.com/auth/youtube.readonly
\item https://www.googleapis.com/auth/yt-analytics-monetary.readonly
\item https://www.googleapis.com/auth/yt-analytics.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/youtube, https://www.googleapis.com/auth/youtube.readonly, https://www.googleapis.com/auth/yt-analytics-monetary.readonly, https://www.googleapis.com/auth/yt-analytics.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/youtube/analytics/}{Google Documentation}
}

