% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtubeAnalytics_objects.R
\name{ResultTable.columnHeaders}
\alias{ResultTable.columnHeaders}
\title{ResultTable.columnHeaders Object}
\usage{
ResultTable.columnHeaders()
}
\value{
ResultTable.columnHeaders object
}
\description{
ResultTable.columnHeaders Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ResultTable functions: \code{\link{ResultTable}}
}

